backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.9.2
pydantic==2.9.2
python-multipart==0.0.12
langchain==0.2.16
langchain-community==0.2.16
sentence-transformers==3.1.1
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
